Fix compilation errors and refactor tensor parallelism
I have addressed several compilation errors that arose from the initial implementation of tensor parallelism. I also refactored the NCCL implementation to better support multiple tensor parallel groups.

Key changes:
- Added `tp_n` to `llama_model_params` and initialized it.
- Fixed various type mismatch and scope errors in `src/llama-model.cpp`.
- Refactored NCCL communicator management in `ggml-cuda.cu` to be more modular and support multiple TP groups.
- Removed global NCCL variables and moved initialization to the split buffer context.
- Corrected the arguments for `nccl_all_reduce` and `nccl_all_gather`.

The code now compiles successfully, but I was unable to test it on a GPU due to environment limitations. The `test-col-tp` test fails with "no CUDA-capable device is detected".

+++ src/llama.cpp
@@ -200,7 +200,7 @@ static struct llama_model * llama_model_load_from_file_impl(
     if (params.tp_n > 1) {
         int n_gpu_groups = 0;
         if (params.tensor_split != nullptr) {
-            for (int i = 0; i < llama_max_devices(); ++i) {
+            for (size_t i = 0; i < llama_max_devices(); ++i) {
                 if (params.tensor_split[i] > 0.0f) {
                     n_gpu_groups++;
                 }
